deposit collateral

+++ src/StableCoin.sol
@@ -20,9 +20,9 @@ contract StableCoin is ERC20Burnable, Ownable {
     error StableCoin__NotZeroAddress();
 
     constructor() ERC20("DecentralizedStableCoin", "DSC") {}
-    
-    function burn(uint _amount) public override onlyOwner {
-        uint balance = balanceOf(msg.sender);
+
+    function burn(uint256 _amount) public override onlyOwner {
+        uint256 balance = balanceOf(msg.sender);
         if (_amount <= 0) {
             revert StableCoin__MustBeGreaterThanZero();
         }
@@ -32,7 +32,7 @@ contract StableCoin is ERC20Burnable, Ownable {
         super.burn(_amount);
     }
 
-    function mint(address _to, uint _amount) external onlyOwner returns(bool) {
+    function mint(address _to, uint256 _amount) external onlyOwner returns (bool) {
         if (_to == address(0)) {
             revert StableCoin__NotZeroAddress();
         }
@@ -42,5 +42,4 @@ contract StableCoin is ERC20Burnable, Ownable {
         _mint(_to, _amount);
         return true;
     }
-
 }
